refactor(task-graph): extract StreamPump from TaskGraphRunner
Moves streaming methods (taskNeedsAccumulation, awaitStreamInputs,
runStreamingTask, createStreamFromTaskEvents, pushStreamToEdges, isPortDelta)
into a new internal class. Also extracts the inline tee block from runTask
into a new prepareStreamingInputs method so the streaming-input handoff is
a named operation rather than buried logic.

runStreamingTask takes a callback for the STREAMING-status edge push so this
task can land before pushStatusFromNodeToEdges itself moves to RunScheduler
in the following commit. The callback is wired by the facade.

awaitStreamInputs takes the ServiceRegistry as a method argument rather than
holding a back-ref or stale constructor-injected reference — lets the facade
pass the current value at call time and keeps StreamPump's per-call shape
explicit.

Pure refactor: no behavior change, no public API change. 1,663 integration
tests pass unchanged, including the streaming-specific suites.
